+++ packages/terra-action-footer/tests/wdio/theme/action-footer-theme-spec.js
@@ -1,0 +1,23 @@
+const {
+  viewports,
+  withActionsThemeables,
+  withoutActionsThemeables,
+} = require('../common');
+
+describe('ActionFooter', () => {
+  describe('Empty', () => {
+    beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/empty-action-footer'));
+
+    Terra.should.beAccessible({ viewports });
+    Terra.should.matchScreenshot({ viewports });
+    Terra.should.themeEachCustomProperty(withoutActionsThemeables);
+  });
+
+  describe('Multiple Start and End Actions', () => {
+    beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/multiple-start-end-action-footer'));
+
+    Terra.should.beAccessible({ viewports });
+    Terra.should.matchScreenshot({ viewports });
+    Terra.should.themeEachCustomProperty(withActionsThemeables);
+  });
+});