+++ packages/terra-action-footer/tests/wdio/theme/centered-action-footer-theme-spec.js
@@ -1,0 +1,23 @@
+const {
+  viewports,
+  withActionsThemeables,
+  withoutActionsThemeables,
+} = require('../common');
+
+describe('CenteredActionFooter', () => {
+  describe('Empty', () => {
+    beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/empty-centered-action-footer'));
+
+    Terra.should.beAccessible({ viewports });
+    Terra.should.matchScreenshot({ viewports });
+    Terra.should.themeEachCustomProperty(withoutActionsThemeables);
+  });
+
+  describe('Multiple Actions', () => {
+    beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/multiple-action-centered-action-footer'));
+
+    Terra.should.beAccessible({ viewports });
+    Terra.should.matchScreenshot({ viewports });
+    Terra.should.themeEachCustomProperty(withActionsThemeables);
+  });
+});